Update Spanish strings
Sorry for the absence, had nougat and no xposed :P, voice call button replacement doesnt work on spanish and its the only use i have for the app, haha. I translated the new string "Voice call" to the toast that appears when you keep the call button pressed "Llamada de voz" like we did before, we have to test it.